+++ src/classic/class/ReactClass.js
