+++ src/classic/class/ReactDoNotBindDeprecated.js
