+++ src/classic/class/__tests__/ReactBind-test.js
