+++ src/classic/propTypes/ReactPropTypeLocations.js
