+++ src/core/__tests__/ReactCompositeComponent-test.js
@@ -85,31 +85,6 @@ describe('ReactCompositeComponent', function() {
     console.warn = warn;
   });
 
-  it('should give context for PropType errors in nested components.', () => {
-    // In this test, we're making sure that if a proptype error is found in a
-    // component, we give a small hint as to which parent instantiated that
-    // component as per warnings about key usage in ReactElementValidator.
-    spyOn(console, 'warn');
-    var MyComp = React.createClass({
-      propTypes: {
-        color: ReactPropTypes.string
-      },
-      render: function() {
-        return <div>My color is {this.color}</div>;
-      }
-    });
-    var ParentComp = React.createClass({
-      render: function() {
-        return <MyComp color={123} />;
-      }
-    });
-    ReactTestUtils.renderIntoDocument(<ParentComp />);
-    expect(console.warn.calls[0].args[0]).toBe(
-      'Warning: Invalid prop `color` of type `number` supplied to `MyComp`, ' +
-      'expected `string`. Check the render method of `ParentComp`.'
-    );
-  });
-
   it('should support rendering to different child types over time', function() {
     var instance = <MorphingComponent />;
     instance = ReactTestUtils.renderIntoDocument(instance);
@@ -299,152 +274,6 @@ describe('ReactCompositeComponent', function() {
     expect(inputProps.prop).not.toBeDefined();
   });
 
-  it('should use default prop value when removing a prop', function() {
-    var Component = React.createClass({
-      getDefaultProps: function() {
-        return {fruit: 'persimmon'};
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-
-    var container = document.createElement('div');
-    var instance = React.render(
-      <Component fruit="mango" />,
-      container
-    );
-    expect(instance.props.fruit).toBe('mango');
-
-    React.render(<Component />, container);
-    expect(instance.props.fruit).toBe('persimmon');
-  });
-
-  it('should normalize props with default values', function() {
-    var Component = React.createClass({
-      propTypes: {prop: ReactPropTypes.string.isRequired},
-      getDefaultProps: function() {
-        return {prop: 'testKey'};
-      },
-      getInitialState: function() {
-        return {prop: this.props.prop + 'State'};
-      },
-      render: function() {
-        return <span>{this.props.prop}</span>;
-      }
-    });
-
-    var instance = ReactTestUtils.renderIntoDocument(<Component />);
-    reactComponentExpect(instance).scalarPropsEqual({prop: 'testKey'});
-    reactComponentExpect(instance).scalarStateEqual({prop: 'testKeyState'});
-
-    ReactTestUtils.renderIntoDocument(<Component prop={null} />);
-
-    expect(console.warn.mock.calls.length).toBe(1);
-    expect(console.warn.mock.calls[0][0]).toBe(
-      'Warning: Required prop `prop` was not specified in `Component`.'
-    );
-  });
-
-  it('should check default prop values', function() {
-    var Component = React.createClass({
-      propTypes: {prop: ReactPropTypes.string.isRequired},
-      getDefaultProps: function() {
-        return {prop: null};
-      },
-      render: function() {
-        return <span>{this.props.prop}</span>;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<Component />);
-
-    expect(console.warn.mock.calls.length).toBe(1);
-    expect(console.warn.mock.calls[0][0]).toBe(
-      'Warning: Required prop `prop` was not specified in `Component`.'
-    );
-  });
-
-  it('should check declared prop types', function() {
-    var Component = React.createClass({
-      propTypes: {
-        prop: ReactPropTypes.string.isRequired
-      },
-      render: function() {
-        return <span>{this.props.prop}</span>;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<Component />);
-    ReactTestUtils.renderIntoDocument(<Component prop={42} />);
-
-    expect(console.warn.mock.calls.length).toBe(2);
-    expect(console.warn.mock.calls[0][0]).toBe(
-      'Warning: Required prop `prop` was not specified in `Component`.'
-    );
-
-    expect(console.warn.mock.calls[1][0]).toBe(
-      'Warning: Invalid prop `prop` of type `number` supplied to ' +
-      '`Component`, expected `string`.'
-    );
-
-    ReactTestUtils.renderIntoDocument(<Component prop="string" />);
-
-    // Should not error for strings
-    expect(console.warn.mock.calls.length).toBe(2);
-  });
-
-  it('should throw on invalid prop types', function() {
-    expect(function() {
-      React.createClass({
-        displayName: 'Component',
-        propTypes: {
-          prop: null
-        },
-        render: function() {
-          return <span>{this.props.prop}</span>;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: Component: prop type `prop` is invalid; ' +
-      'it must be a function, usually from React.PropTypes.'
-    );
-  });
-
-  it('should throw on invalid context types', function() {
-    expect(function() {
-      React.createClass({
-        displayName: 'Component',
-        contextTypes: {
-          prop: null
-        },
-        render: function() {
-          return <span>{this.props.prop}</span>;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: Component: context type `prop` is invalid; ' +
-      'it must be a function, usually from React.PropTypes.'
-    );
-  });
-
-  it('should throw on invalid child context types', function() {
-    expect(function() {
-      React.createClass({
-        displayName: 'Component',
-        childContextTypes: {
-          prop: null
-        },
-        render: function() {
-          return <span>{this.props.prop}</span>;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: Component: child context type `prop` is invalid; ' +
-      'it must be a function, usually from React.PropTypes.'
-    );
-  });
-
   it('should not allow `forceUpdate` on unmounted components', function() {
     var container = document.createElement('div');
     document.documentElement.appendChild(container);
@@ -574,228 +403,6 @@ describe('ReactCompositeComponent', function() {
     expect(ReactCurrentOwner.current).toBe(null);
   });
 
-  it('should support mixins with getInitialState()', function() {
-    var Mixin = {
-      getInitialState: function() {
-        return {mixin: true};
-      }
-    };
-    var Component = React.createClass({
-      mixins: [Mixin],
-      getInitialState: function() {
-        return {component: true};
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state.component).toBe(true);
-    expect(instance.state.mixin).toBe(true);
-  });
-
-  it('should throw with conflicting getInitialState() methods', function() {
-    var Mixin = {
-      getInitialState: function() {
-        return {x: true};
-      }
-    };
-    var Component = React.createClass({
-      mixins: [Mixin],
-      getInitialState: function() {
-        return {x: true};
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    expect(function() {
-      instance = ReactTestUtils.renderIntoDocument(instance);
-    }).toThrow(
-      'Invariant Violation: mergeIntoWithNoDuplicateKeys(): ' +
-      'Tried to merge two objects with the same key: `x`. This conflict ' +
-      'may be due to a mixin; in particular, this may be caused by two ' +
-      'getInitialState() or getDefaultProps() methods returning objects ' +
-      'with clashing keys.'
-    );
-  });
-
-  it('should not mutate objects returned by getInitialState()', function() {
-    var Mixin = {
-      getInitialState: function() {
-        return Object.freeze({mixin: true});
-      }
-    };
-    var Component = React.createClass({
-      mixins: [Mixin],
-      getInitialState: function() {
-        return Object.freeze({component: true});
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    expect(() => {
-      ReactTestUtils.renderIntoDocument(<Component />);
-    }).not.toThrow();
-  });
-
-  it('should work with object getInitialState() return values', function() {
-    var Component = React.createClass({
-      getInitialState: function() {
-        return {
-          occupation: 'clown'
-        };
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state.occupation).toEqual('clown');
-  });
-
-  it('should throw with non-object getInitialState() return values', function() {
-    [['an array'], 'a string', 1234].forEach(function(state) {
-      var Component = React.createClass({
-        getInitialState: function() {
-          return state;
-        },
-        render: function() {
-          return <span />;
-        }
-      });
-      var instance = <Component />;
-      expect(function() {
-        instance = ReactTestUtils.renderIntoDocument(instance);
-      }).toThrow(
-        'Invariant Violation: Component.getInitialState(): ' +
-        'must return an object or null'
-      );
-    });
-  });
-
-  it('should work with a null getInitialState() return value', function() {
-    var Component = React.createClass({
-      getInitialState: function() {
-        return null;
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    expect(
-      () => ReactTestUtils.renderIntoDocument(<Component />)
-    ).not.toThrow();
-  });
-
-  it('should work with a null getInitialState return value and a mixin', () => {
-    var Component;
-    var instance;
-
-    var Mixin = {
-      getInitialState: function() {
-        return {foo: 'bar'};
-      }
-    };
-    Component = React.createClass({
-      mixins: [Mixin],
-      getInitialState: function() {
-        return null;
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    expect(
-      () => ReactTestUtils.renderIntoDocument(<Component />)
-    ).not.toThrow();
-
-    instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state).toEqual({foo: 'bar'});
-
-    // Also the other way round should work
-    var Mixin2 = {
-      getInitialState: function() {
-        return null;
-      }
-    };
-    Component = React.createClass({
-      mixins: [Mixin2],
-      getInitialState: function() {
-        return {foo: 'bar'};
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    expect(
-      () => ReactTestUtils.renderIntoDocument(<Component />)
-    ).not.toThrow();
-
-    instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state).toEqual({foo: 'bar'});
-
-    // Multiple mixins should be fine too
-    Component = React.createClass({
-      mixins: [Mixin, Mixin2],
-      getInitialState: function() {
-        return {x: true};
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    expect(
-      () => ReactTestUtils.renderIntoDocument(<Component />)
-    ).not.toThrow();
-
-    instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state).toEqual({foo: 'bar', x: true});
-  });
-
-  it('should work with object getInitialState() return values', function() {
-    var Component = React.createClass({
-      getInitialState: function() {
-        return {
-          occupation: 'clown'
-        };
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.state.occupation).toEqual('clown');
-  });
-
-  it('should throw with non-object getInitialState() return values', function() {
-    [['an array'], 'a string', 1234].forEach(function(state) {
-      var Component = React.createClass({
-        getInitialState: function() {
-          return state;
-        },
-        render: function() {
-          return <span />;
-        }
-      });
-      var instance = <Component />;
-      expect(function() {
-        instance = ReactTestUtils.renderIntoDocument(instance);
-      }).toThrow(
-        'Invariant Violation: Component.getInitialState(): ' +
-        'must return an object or null'
-      );
-    });
-  });
-
   it('should call componentWillUnmount before unmounting', function() {
     var container = document.createElement('div');
     var innerUnmounted = false;
@@ -860,88 +467,6 @@ describe('ReactCompositeComponent', function() {
       expect(console.warn.mock.calls[0][0]).toBe(
         'Component.shouldComponentUpdate(): Returned undefined instead of a ' +
         'boolean value. Make sure to return true or false.'
-      );
-    } finally {
-      console.warn = warn;
-    }
-  });
-
-  it('should warn when mispelling shouldComponentUpdate', function() {
-    var warn = console.warn;
-    console.warn = mocks.getMockFunction();
-
-    try {
-      React.createClass({
-        componentShouldUpdate: function() {
-          return false;
-        },
-        render: function() {
-          return <div />;
-        }
-      });
-      expect(console.warn.mock.calls.length).toBe(1);
-      expect(console.warn.mock.calls[0][0]).toBe(
-        'A component has a method called componentShouldUpdate(). Did you ' +
-        'mean shouldComponentUpdate()? The name is phrased as a question ' +
-        'because the function is expected to return a value.'
-      );
-
-      var NamedComponent = React.createClass({
-        componentShouldUpdate: function() {
-          return false;
-        },
-        render: function() {
-          return <div />;
-        }
-      });
-      expect(console.warn.mock.calls.length).toBe(2);
-      expect(console.warn.mock.calls[1][0]).toBe(
-        'NamedComponent has a method called componentShouldUpdate(). Did you ' +
-        'mean shouldComponentUpdate()? The name is phrased as a question ' +
-        'because the function is expected to return a value.'
-      );
-
-      <NamedComponent />; // Shut up lint
-    } finally {
-      console.warn = warn;
-    }
-  });
-
-  xit('should warn when using deprecated non-static spec keys', function() {
-    var warn = console.warn;
-    console.warn = mocks.getMockFunction();
-    try {
-      React.createClass({
-        mixins: [{}],
-        propTypes: {
-          foo: ReactPropTypes.string
-        },
-        contextTypes: {
-          foo: ReactPropTypes.string
-        },
-        childContextTypes: {
-          foo: ReactPropTypes.string
-        },
-        render: function() {
-          return <div />;
-        }
-      });
-      expect(console.warn.mock.calls.length).toBe(4);
-      expect(console.warn.mock.calls[0][0]).toBe(
-        'createClass(...): `mixins` is now a static property and should ' +
-        'be defined inside "statics".'
-      );
-      expect(console.warn.mock.calls[1][0]).toBe(
-        'createClass(...): `propTypes` is now a static property and should ' +
-        'be defined inside "statics".'
-      );
-      expect(console.warn.mock.calls[2][0]).toBe(
-        'createClass(...): `contextTypes` is now a static property and ' +
-        'should be defined inside "statics".'
-      );
-      expect(console.warn.mock.calls[3][0]).toBe(
-        'createClass(...): `childContextTypes` is now a static property and ' +
-        'should be defined inside "statics".'
       );
     } finally {
       console.warn = warn;
@@ -1072,413 +597,6 @@ describe('ReactCompositeComponent', function() {
       '(see: http://fb.me/react-context-by-parent)'
     );
 
-  });
-
-  it('should check context types', function() {
-    var Component = React.createClass({
-      contextTypes: {
-        foo: ReactPropTypes.string.isRequired
-      },
-
-      render: function() {
-        return <div />;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<Component />);
-
-    expect(console.warn.mock.calls.length).toBe(1);
-    expect(console.warn.mock.calls[0][0]).toBe(
-      'Warning: Required context `foo` was not specified in `Component`.'
-    );
-
-    var ComponentInFooStringContext = React.createClass({
-      childContextTypes: {
-        foo: ReactPropTypes.string
-      },
-
-      getChildContext: function() {
-        return {
-          foo: this.props.fooValue
-        };
-      },
-
-      render: function() {
-        return <Component />;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<ComponentInFooStringContext fooValue={'bar'} />);
-
-    // Previous call should not error
-    expect(console.warn.mock.calls.length).toBe(1);
-
-    var ComponentInFooNumberContext = React.createClass({
-      childContextTypes: {
-        foo: ReactPropTypes.number
-      },
-
-      getChildContext: function() {
-        return {
-          foo: this.props.fooValue
-        };
-      },
-
-      render: function() {
-        return <Component />;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<ComponentInFooNumberContext fooValue={123} />);
-
-    expect(console.warn.mock.calls.length).toBe(2);
-    expect(console.warn.mock.calls[1][0]).toBe(
-      'Warning: Invalid context `foo` of type `number` supplied ' +
-      'to `Component`, expected `string`.' +
-      ' Check the render method of `ComponentInFooNumberContext`.'
-    );
-  });
-
-  it('should check child context types', function() {
-    var Component = React.createClass({
-      childContextTypes: {
-        foo: ReactPropTypes.string.isRequired,
-        bar: ReactPropTypes.number
-      },
-
-      getChildContext: function() {
-        return this.props.testContext;
-      },
-
-      render: function() {
-        return <div />;
-      }
-    });
-
-    ReactTestUtils.renderIntoDocument(<Component testContext={{bar: 123}} />);
-    expect(console.warn.mock.calls.length).toBe(2);
-    expect(console.warn.mock.calls[0][0]).toBe(
-      'Warning: Required child context `foo` was not specified in `Component`.'
-    );
-    expect(console.warn.mock.calls[1][0]).toBe(
-      'Warning: Required child context `foo` was not specified in `Component`.'
-    );
-
-    ReactTestUtils.renderIntoDocument(<Component testContext={{foo: 123}} />);
-
-    expect(console.warn.mock.calls.length).toBe(4);
-    expect(console.warn.mock.calls[3][0]).toBe(
-      'Warning: Invalid child context `foo` of type `number` ' +
-      'supplied to `Component`, expected `string`.'
-    );
-
-    ReactTestUtils.renderIntoDocument(
-      <Component testContext={{foo: 'foo', bar: 123}} />
-    );
-
-    ReactTestUtils.renderIntoDocument(
-      <Component testContext={{foo: 'foo'}} />
-    );
-
-    // Previous calls should not log errors
-    expect(console.warn.mock.calls.length).toBe(4);
-  });
-
-  it('should filter out context not in contextTypes', function() {
-    var Component = React.createClass({
-      contextTypes: {
-        foo: ReactPropTypes.string
-      },
-
-      render: function() {
-        return <div />;
-      }
-    });
-
-    var ComponentInFooBarContext = React.createClass({
-      childContextTypes: {
-        foo: ReactPropTypes.string,
-        bar: ReactPropTypes.number
-      },
-
-      getChildContext: function() {
-        return {
-          foo: 'abc',
-          bar: 123
-        };
-      },
-
-      render: function() {
-        return <Component />;
-      }
-    });
-
-    var instance = ReactTestUtils.renderIntoDocument(<ComponentInFooBarContext />);
-    reactComponentExpect(instance).expectRenderedChild().scalarContextEqual({foo: 'abc'});
-  });
-
-  it('should filter context properly in callbacks', function() {
-    var actualComponentWillReceiveProps;
-    var actualShouldComponentUpdate;
-    var actualComponentWillUpdate;
-    var actualComponentDidUpdate;
-
-    var Parent = React.createClass({
-      childContextTypes: {
-        foo: ReactPropTypes.string.isRequired,
-        bar: ReactPropTypes.string.isRequired
-      },
-
-      getChildContext: function() {
-        return {
-          foo: this.props.foo,
-          bar: "bar"
-        };
-      },
-
-      render: function() {
-        return <Component />;
-      }
-    });
-
-    var Component = React.createClass({
-      contextTypes: {
-        foo: ReactPropTypes.string
-      },
-
-      componentWillReceiveProps: function(nextProps, nextContext) {
-        actualComponentWillReceiveProps = nextContext;
-        return true;
-      },
-
-      shouldComponentUpdate: function(nextProps, nextState, nextContext) {
-        actualShouldComponentUpdate = nextContext;
-        return true;
-      },
-
-      componentWillUpdate: function(nextProps, nextState, nextContext) {
-        actualComponentWillUpdate = nextContext;
-      },
-
-      componentDidUpdate: function(prevProps, prevState, prevContext) {
-        actualComponentDidUpdate = prevContext;
-      },
-
-      render: function() {
-        return <div />;
-      }
-    });
-
-    var instance = <Parent foo="abc" />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    instance.replaceProps({foo: "def"});
-    expect(actualComponentWillReceiveProps).toEqual({foo: 'def'});
-    expect(actualShouldComponentUpdate).toEqual({foo: 'def'});
-    expect(actualComponentWillUpdate).toEqual({foo: 'def'});
-    expect(actualComponentDidUpdate).toEqual({foo: 'abc'});
-  });
-
-  it('should support statics', function() {
-    var Component = React.createClass({
-      statics: {
-        abc: 'def',
-        def: 0,
-        ghi: null,
-        jkl: 'mno',
-        pqr: function() {
-          return this;
-        }
-      },
-
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.constructor.abc).toBe('def');
-    expect(Component.abc).toBe('def');
-    expect(instance.constructor.def).toBe(0);
-    expect(Component.def).toBe(0);
-    expect(instance.constructor.ghi).toBe(null);
-    expect(Component.ghi).toBe(null);
-    expect(instance.constructor.jkl).toBe('mno');
-    expect(Component.jkl).toBe('mno');
-    expect(instance.constructor.pqr()).toBe(Component);
-    expect(Component.pqr()).toBe(Component);
-  });
-
-  it('should throw if a reserved property is in statics', function() {
-    expect(function() {
-      React.createClass({
-        statics: {
-          getDefaultProps: function() {
-            return {
-              foo: 0
-            };
-          }
-        },
-
-        render: function() {
-          return <span />;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: ReactClass: You are attempting to ' +
-      'define a reserved property, `getDefaultProps`, that shouldn\'t be on ' +
-      'the "statics" key. Define it as an instance property instead; it ' +
-      'will still be accessible on the constructor.'
-    );
-  });
-
-  it('should support statics in mixins', function() {
-    var Mixin = {
-      statics: {
-        foo: 'bar'
-      }
-    };
-    var Component = React.createClass({
-      mixins: [Mixin],
-
-      statics: {
-        abc: 'def'
-      },
-
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-    expect(instance.constructor.foo).toBe('bar');
-    expect(Component.foo).toBe('bar');
-    expect(instance.constructor.abc).toBe('def');
-    expect(Component.abc).toBe('def');
-  });
-
-  it("should throw if mixins override each others' statics", function() {
-    expect(function() {
-      var Mixin = {
-        statics: {
-          abc: 'foo'
-        }
-      };
-      React.createClass({
-        mixins: [Mixin],
-
-        statics: {
-          abc: 'bar'
-        },
-
-        render: function() {
-          return <span />;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: ReactClass: You are attempting to ' +
-      'define `abc` on your component more than once. This conflict may be ' +
-      'due to a mixin.'
-    );
-  });
-
-  it("should throw if mixins override functions in statics", function() {
-    expect(function() {
-      var Mixin = {
-        statics: {
-          abc: function() { console.log('foo'); }
-        }
-      };
-      React.createClass({
-        mixins: [Mixin],
-
-        statics: {
-          abc: function() { console.log('bar'); }
-        },
-
-        render: function() {
-          return <span />;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: ReactClass: You are attempting to ' +
-      'define `abc` on your component more than once. This conflict may be ' +
-      'due to a mixin.'
-    );
-  });
-
-  it("should throw if the mixin is a React component", function() {
-    expect(function() {
-      React.createClass({
-        mixins: [<div />],
-
-        render: function() {
-          return <span />;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: ReactClass: You\'re attempting to ' +
-      'use a component as a mixin. Instead, just use a regular object.'
-    );
-  });
-
-  it("should throw if the mixin is a React component class", function() {
-    expect(function() {
-      var Component = React.createClass({
-        render: function() {
-          return <span />;
-        }
-      });
-
-      React.createClass({
-        mixins: [Component],
-
-        render: function() {
-          return <span />;
-        }
-      });
-    }).toThrow(
-      'Invariant Violation: ReactClass: You\'re attempting to ' +
-      'use a component class as a mixin. Instead, just use a regular object.'
-    );
-  });
-
-  it('should have bound the mixin methods to the component', function() {
-    var mixin = {
-      mixinFunc: function() {return this;}
-    };
-
-    var Component = React.createClass({
-      mixins: [mixin],
-      componentDidMount: function() {
-        expect(this.mixinFunc()).toBe(this);
-      },
-      render: function() {
-        return <span />;
-      }
-    });
-    var instance = <Component />;
-    instance = ReactTestUtils.renderIntoDocument(instance);
-  });
-
-  it('should include the mixin keys in even if their values are falsy',
-    function() {
-      var mixin = {
-        keyWithNullValue: null,
-        randomCounter: 0
-      };
-
-      var Component = React.createClass({
-        mixins: [mixin],
-        componentDidMount: function() {
-          expect(this.randomCounter).toBe(0);
-          expect(this.keyWithNullValue).toBeNull();
-        },
-        render: function() {
-          return <span />;
-        }
-      });
-      var instance = <Component />;
-      instance = ReactTestUtils.renderIntoDocument(instance);
   });
 
   it('should disallow nested render calls', function() {
